✨ Add a NoOp store
